Various changes:
(i) Added function to dust/utils.py to return dust-free name for a dataset (i.e. function to remove dust extension from nanme).

(ii) Restructuring of CharlotFall2000 class in dust/CharlotFall2000.py so that can work with 'total' datasets.

(iii) Restructuring of derivedProperties class in derivedProperties.py and addition of a new class function that checks the nature of a single property and calls the appropriate function. Includes ability to skip dust attenuation for selected properties.